Address nits.
