+++ src/librustc_typeck/check/assoc.rs
@@ -80,11 +80,9 @@ impl<'a,'tcx> TypeFolder<'tcx> for AssociatedTypeNormalizer<'a,'tcx> {
                         self.span,
                         self.body_id,
                         ObligationCauseCode::MiscObligation);
-                let trait_ref = data.trait_ref.clone();
                 self.fulfillment_cx
-                    .normalize_associated_type(self.infcx,
-                                               trait_ref,
-                                               data.item_name,
+                    .normalize_projection_type(self.infcx,
+                                               data.clone(),
                                                cause)
             }
             _ => {
